requirements.txt
absl-py==1.4.0
action-tutorials-py==0.9.3
ament-copyright==0.9.6
ament-cppcheck==0.9.6
ament-cpplint==0.9.6
ament-flake8==0.9.6
ament-index-python==1.1.0
ament-lint==0.9.6
ament-lint-cmake==0.9.6
ament-package==0.9.5
ament-pep257==0.9.6
ament-uncrustify==0.9.6
ament-xmllint==0.9.6
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
demo-nodes-py==0.9.3
domain-coordinator==0.9.2
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
flake8==6.0.0
fonttools==4.39.2
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
launch==0.10.6
launch-ros==0.11.4
launch-testing==0.10.6
launch-testing-ros==0.11.4
launch-xml==0.10.6
launch-yaml==0.10.6
Markdown==3.4.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
osrf-pycommon==0.1.11
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
quality-of-service-demo-py==0.9.3
requests==2.28.2
requests-oauthlib==1.3.1
ros2action==0.9.10
ros2bag==0.3.8
ros2cli==0.9.10
ros2component==0.9.10
ros2doctor==0.9.10
ros2interface==0.9.10
ros2launch==0.11.4
ros2lifecycle==0.9.10
ros2multicast==0.9.10
ros2node==0.9.10
ros2param==0.9.10
ros2pkg==0.9.10
ros2run==0.9.10
ros2service==0.9.10
ros2topic==0.9.10
rosidl-runtime-py==0.9.1
rpyutils==0.2.0
rqt-action==0.4.9
rqt-graph==1.1.2
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-msg==1.0.5
rqt-plot==1.1.0
rqt-publisher==1.1.3
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-top==1.0.2
rqt-topic==1.2.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sros2==0.9.4
teleop-twist-keyboard==2.3.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tomli==2.0.1
topic-monitor==0.9.3
torch==1.12.0+cu113
torchsummary==1.5.1
torchvision==0.13.0+cu113
types-PyYAML==6.0.12.8
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
